Separando os cards de incentivo e investimento em dois componentes distintos.
